Add GiveOrnament && GiveTitle effect. Fix NpcPrestige Message if level < 200

+++ trunk/Server/Stump.Server.WorldServer/Handlers/Inventory/InventoryItemsHandler.cs
@@ -151,7 +151,7 @@ namespace Stump.Server.WorldServer.Handlers.Inventory
 
         public static void SendInventoryWeightMessage(WorldClient client)
         {
-            client.Send(new InventoryWeightMessage((int) client.Character.Inventory.Weight,
+            client.Send(new InventoryWeightMessage(client.Character.Inventory.Weight,
                                                    (int) client.Character.Inventory.WeightTotal));
         }
